attempted to correct all the file structures added pages, styles, components, directories added _app.js for global.css, which is required for next. need to fix tailwind css, so it can be global
