+++ tastyfoodapp/pages/_app.js
@@ -1,0 +1,5 @@
+import '../styles/global.css';
+
+export default function App({ Component, pageProps }) {
+    return <Component {...pageProps} />
+}
