--- tastyfoodapp/src/app/layout.js
@@ -1,23 +1,0 @@
-import { Inter } from "next/font/google";
-import "./globals.css";
-import Navbar from "@/components/navbar";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-const RootLayout = ({ children }) => {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <Navbar/>
-        {children}
-      </body>
-    </html>
-  );
-};
-
-export default RootLayout;
